+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/internal/EventImpl.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.Objects;
 import org.hiero.base.Clearable;
 import org.hiero.consensus.model.event.EventDescriptorWrapper;
+import org.hiero.consensus.model.event.LinkedEvent;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.hashgraph.ConsensusConstants;
 import org.hiero.consensus.model.node.NodeId;
